backend/requirements/base.txt
# Core dependencies
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
python-dotenv==1.0
celery==5.4.0

